mux: fix FW build with xt-clang
Newer Xtensa toolchain replaces xt-xcc with xt-clang.
For some reason xt-clang fails to build SOF when uint8_t type
is used as index of for() loop. This change unblocks building of
SOF with xt-clang.

+++ src/audio/mux/mux.c
@@ -97,8 +97,8 @@ static bool mux_mix_check(struct sof_mux_config *cfg)
 static int mux_set_values(struct comp_dev *dev, struct comp_data *cd,
 			  struct sof_mux_config *cfg)
 {
-	uint8_t i;
-	uint8_t j;
+	int i;
+	int j;
 
 	comp_info(dev, "mux_set_values()");
